+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# v1.5.10
+## 09/29/2021
+
+1. [](#improved)
+    * Improved `UniformResourceLocator` to support `file://` scheme
+1. [](#bugfix)
+    * Fixed blueprint merge where second blueprint has non-array field definition
+    * Fixed implicit cast from null to string in `UniformResourceLocator`
+
 # v1.5.9
 ## 04/14/2021